+++ assets/Assets.php
@@ -15,6 +15,14 @@ use yii\web\AssetBundle;
 
 class Assets extends AssetBundle
 {
+    /**
+     * v1.5 compatibility defer script loading
+     *
+     * Migrate to HumHub AssetBundle once minVersion is >=1.5
+     *
+     * @var bool
+     */
+    public $defer = true;
 
     public $publishOptions = ['forceCopy' => false];